[#648] Separete deprecated server-side views
Moved deprected server-side views from main views file to
views/deprecated.rb where they can be chopped off at once, together (no
template left behind).

* Ensures clearer separation between active and deprecated code.
* Aids in future code removals and updates.
* No functional changes or disruptions to existing features.
